fix bug load high pieceId which has pieceSize * pieceIdx > Integer.MAX_VALUE

+++ src/main/java/com/solt/media/stream/helper/MdDataHelper.java
@@ -51,7 +51,7 @@ public class MdDataHelper implements TDataHelper {
 			return Result.ERROR_RESULT;
 		}
 		
-		long startBytes = pieceIdx * pieceSize - itemOffset;
+		long startBytes = ((long)pieceIdx) * pieceSize - itemOffset;
 		long endBytes = startBytes + pieceSize;
 		int state = Result.COMPLETE;
 		int offset = 0;
@@ -82,7 +82,7 @@ public class MdDataHelper implements TDataHelper {
 			return false;
 		}
 		
-		long startBytes = pieceIdx * pieceSize - itemOffset;
+		long startBytes = ((long)pieceIdx) * pieceSize - itemOffset;
 		long endBytes = startBytes + pieceSize;
 		if (startBytes < 0 || endBytes > itemLength) {
 			return false;
@@ -134,7 +134,7 @@ public class MdDataHelper implements TDataHelper {
 			return false;
 		}
 		
-		long startBytes = pieceIdx * pieceSize - itemOffset;
+		long startBytes = ((long)pieceIdx) * pieceSize - itemOffset;
 		long endBytes = startBytes + pieceSize;
 		if (startBytes < 0 || endBytes > itemLength) {
 			return false;
